# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A tunnel service for exposing local services to the internet",
        "title": "ShipIt Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ShipIt Support",
            "url": "http://www.shipit.com/support",
            "email": "support@shipit.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/analytics/overview": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get user's analytics overview for the specified period",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get analytics overview",
                "parameters": [
                    {
                        "type": "string",
                        "default": "24h",
                        "description": "Time period (24h, 7d, 30d)",
                        "name": "period",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Analytics overview",
                        "schema": {
                            "$ref": "#/definitions/handlers.AnalyticsOverviewResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid period parameter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/analytics/traffic": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed traffic analytics including top visitors, paths, status codes, and geographic data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get traffic analytics",
                "parameters": [
                    {
                        "type": "string",
                        "default": "24h",
                        "description": "Time period for analytics",
                        "name": "period",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Traffic analytics",
                        "schema": {
                            "$ref": "#/definitions/handlers.TrafficAnalyticsResponse"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/analytics/tunnels/{tunnel_id}/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed analytics for a specific tunnel including metrics and time series data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get tunnel analytics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tunnel ID",
                        "name": "tunnel_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "24h",
                        "description": "Time period for analytics",
                        "name": "period",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "requests,bandwidth,latency",
                        "description": "Metrics to include",
                        "name": "metrics",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tunnel analytics",
                        "schema": {
                            "$ref": "#/definitions/handlers.TunnelAnalyticsResponse"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Tunnel not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.AnalyticsOverviewResponse": {
            "type": "object",
            "properties": {
                "active_tunnels": {
                    "type": "integer"
                },
                "generated_at": {
                    "type": "string"
                },
                "period": {
                    "type": "string"
                },
                "total_bandwidth": {
                    "description": "Human readable (e.g., \"2.3GB\")",
                    "type": "string"
                },
                "total_bandwidth_bytes": {
                    "type": "integer"
                },
                "total_requests": {
                    "type": "integer"
                },
                "total_tunnels": {
                    "type": "integer"
                }
            }
        },
        "handlers.CountryStat": {
            "type": "object",
            "properties": {
                "country_code": {
                    "type": "string"
                },
                "country_name": {
                    "type": "string"
                },
                "percent": {
                    "type": "number"
                },
                "requests": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.PathStat": {
            "type": "object",
            "properties": {
                "bandwidth_bytes": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "requests": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.TrafficAnalyticsResponse": {
            "type": "object",
            "properties": {
                "countries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.CountryStat"
                    }
                },
                "generated_at": {
                    "type": "string"
                },
                "period": {
                    "type": "string"
                },
                "status_codes": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "top_paths": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.PathStat"
                    }
                },
                "top_visitors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.VisitorStat"
                    }
                },
                "user_agents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.UserAgentStat"
                    }
                }
            }
        },
        "handlers.TunnelAnalyticsPoint": {
            "type": "object",
            "properties": {
                "bytes_in": {
                    "type": "integer"
                },
                "bytes_out": {
                    "type": "integer"
                },
                "error_count": {
                    "type": "integer"
                },
                "requests_count": {
                    "type": "integer"
                },
                "response_time_avg": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "handlers.TunnelAnalyticsResponse": {
            "type": "object",
            "properties": {
                "metrics": {
                    "$ref": "#/definitions/handlers.TunnelMetricsSummary"
                },
                "period": {
                    "type": "string"
                },
                "time_series": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.TunnelAnalyticsPoint"
                    }
                },
                "tunnel_id": {
                    "type": "string"
                }
            }
        },
...
        "handlers.TunnelMetricsSummary": {
            "type": "object",
            "properties": {
                "average_latency_ms": {
                    "type": "number"
                },
                "error_rate_percent": {
                    "type": "number"
                },
                "total_bandwidth": {
                    "type": "string"
                },
                "total_bandwidth_bytes": {
                    "type": "integer"
                },
                "total_requests": {
                    "type": "integer"
                },
                "uptime_percent": {
                    "type": "number"
                }
            }
        },
...
        "handlers.UserAgentStat": {
            "type": "object",
            "properties": {
                "percent": {
                    "type": "number"
                },
                "requests": {
                    "type": "integer"
                },
                "user_agent": {
                    "type": "string"
                }
            }
        },
...
        "handlers.VisitorStat": {
            "type": "object",
            "properties": {
                "bandwidth_bytes": {
                    "type": "integer"
                },
                "country": {
                    "type": "string"
                },
                "ip_address": {
                    "type": "string"
                },
                "requests": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key authentication for CLI agents and web users",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT token authentication for web users",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}